requirements.txt
langchain
langchain-groq
langchain-community
langchain-core
python-dotenv
